refactor: 도서 목록 페이지 레이아웃 컴포넌트명 변경
