+++ src/app/books/[list]/_components/BookListPageLayout/index.module.scss
